- update from HEAD

+++ lustre/tests/createtest.c
@@ -96,9 +96,9 @@ int main(int argc, char *argv[])
 			if (i == S_IFCHR || i == S_IFBLK) {
 				if (st.st_rdev != 0x1234) {
 					fprintf(stderr, "%s: ERROR rdev %s: "
-					        "%lu != 0x1234",
+					        "%llu != 0x1234",
 					        argv[0], name,
-                                                (unsigned long) st.st_rdev);
+					        (unsigned long long)st.st_rdev);
 					exit(13);
 				}
 			}
